Add CTR
Adding new mode of operation : Counter Mode. The initial nonce/counter is obtained by generating a number with the LFSR based PRNG, and then using a function to increment each element of the nonce. The substitution-permutation network of AES will offer a bit avalanche efficient enough for the single-bit changes of the nonce to actually provide cryptographic security.